+++ unit_tests/library/route.cpp
@@ -7,6 +7,7 @@
 
 #include "osrm/coordinate.hpp"
 #include "osrm/engine_config.hpp"
+#include "osrm/exception.hpp"
 #include "osrm/json_container.hpp"
 #include "osrm/json_container.hpp"
 #include "osrm/osrm.hpp"